feat: separate bundle for edge envs
