+++ examples/next.js-typescript/lib/session.ts
@@ -16,3 +16,9 @@ declare module "iron-session" {
     user?: User;
   }
 }
+
+declare module "iron-session/edge" {
+  interface IronSessionData {
+    user?: User;
+  }
+}
