+++ package.json
@@ -32,6 +32,11 @@
       "import": "./express/dist/index.mjs",
       "require": "./express/dist/index.js",
       "types": "./express/dist/index.d.ts"
+    },
+    "./edge": {
+      "import": "./edge/dist/index.mjs",
+      "require": "./edge/dist/index.js",
+      "types": "./edge/dist/index.d.ts"
     }
   },
   "main": "./dist/index.js",
@@ -40,20 +45,23 @@
   "files": [
     "dist/",
     "next/",
-    "express/"
+    "express/",
+    "edge/"
   ],
   "scripts": {
-    "build": "npm run build:clean && npm run build:src && npm run build:next && npm run build:express",
-    "build:clean": "rimraf dist/ next/dist express/dist",
+    "build": "npm run build:clean && npm run build:src && npm run build:next && npm run build:express && npm run build:edge",
+    "build:clean": "rimraf dist/ next/dist express/dist edge/dist",
+    "build:edge": "tsup edge/index.ts -d edge/dist",
     "build:express": "tsup express/index.ts -d express/dist",
     "build:next": "tsup next/index.ts -d next/dist",
     "build:src": "tsup src/index.ts",
     "installExamples": "(cd examples/express && npm ci) && (cd examples/next.js && npm ci) && (cd examples/next.js-typescript && npm ci)",
-    "lint": "eslint --ext ts,js,jsx,tsx src/ next/ express/ examples/",
-    "prepublishOnly": "npm run build && cp next/dist/* next/ && cp express/dist/* express/",
-    "postpublish": "rm next/*.d.ts next/*.js next/*.map next/*.mjs && rm express/*.d.ts express/*.js express/*.map express/*.mjs",
+    "lint": "eslint --ext ts,js,jsx,tsx src/ next/ express/ edge/ examples/",
+    "prepublishOnly": "npm run build && cp next/dist/* next/ && cp express/dist/* express/ && cp edge/dist/* edge/",
+    "postpublish": "rm next/*.d.ts next/*.js next/*.map next/*.mjs && rm express/*.d.ts express/*.js express/*.map express/*.mjs && rm edge/*.d.ts edge/*.js edge/*.map edge/*.mjs",
     "test": "jest --coverage --ci && npm run installExamples && npm run lint && tsc --noEmit && npm run build && npm run build:clean",
     "watch": "npm run build:clean && concurrently \"npm:watch:*\"",
+    "watch:edge": "tsup edge/index.ts -d edge/dist --watch edge/index.ts",
     "watch:express": "tsup express/index.ts -d express/dist --watch express/index.ts",
     "watch:next": "tsup next/index.ts -d next/dist --watch next/index.ts",
     "watch:src": "tsup src/index.ts --watch src/index.ts"
@@ -121,26 +129,27 @@
     }
   },
   "dependencies": {
+    "@peculiar/webcrypto": "^1.4.0",
     "@types/cookie": "^0.5.1",
     "@types/express": "^4.17.13",
-    "@types/node": "^16.11.7",
+    "@types/node": "^17.0.41",
     "cookie": "^0.5.0",
-    "iron-webcrypto": "^0.0.1"
+    "iron-webcrypto": "^0.1.0"
   },
   "devDependencies": {
-    "@swc/core": "1.2.171",
-    "@swc/jest": "0.2.20",
+    "@swc/core": "1.2.197",
+    "@swc/jest": "0.2.21",
     "@tsconfig/node12": "1.0.9",
-    "@types/jest": "27.4.1",
-    "@typescript-eslint/eslint-plugin": "5.20.0",
-    "@typescript-eslint/parser": "5.20.0",
-    "concurrently": "7.1.0",
-    "eslint": "8.14.0",
-    "jest": "27.5.1",
+    "@types/jest": "28.1.1",
+    "@typescript-eslint/eslint-plugin": "5.27.1",
+    "@typescript-eslint/parser": "5.27.1",
+    "concurrently": "7.2.1",
+    "eslint": "8.17.0",
+    "jest": "28.1.1",
     "prettier": "2.6.2",
-    "prettier-plugin-packagejson": "2.2.17",
+    "prettier-plugin-packagejson": "2.2.18",
     "rimraf": "3.0.2",
-    "tsup": "5.12.6",
+    "tsup": "6.1.0",
     "typescript": "4.7.3"
   },
   "peerDependencies": {
